都道府県・現在地RxSwift化

+++ TenkiAppPart2/Models/Renponse/OneCallResponse.swift
@@ -39,7 +39,7 @@ struct Current: Codable {
 
 struct Minutely: Codable {
     let dt: Int
-    let precipitation: Int
+    let precipitation: Double
 }
 
 struct Hourly: Codable {
